NXP-25633: Refactor CSVExport Action
