+++ nuxeo-runtime/nuxeo-stream/src/main/resources/tools-log4j.xml
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 <?xml version="1.0" encoding="UTF-8"?>
 <Configuration>
   <Appenders>
